pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.22.0"
Pillow = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
